Removed sudo package, added new mixer module
